programmable_rss: split out single queue case into separate function
Currently, the single queue case still prints all of the configured
queues for the NIC, even though only one is active. Instead of mucking
around with poll_stats() and the indirection map for multiple cases,
just split the single queue case off into a separate function. This
reuses the indirection map as a map to keep track of its stats
(purely out of convenience).